+++ ui/pages/join.tsx
@@ -14,7 +14,7 @@ import PassportExpiration from '../components/PassportExpiration'
 import { balancerDomain, nationToken, veNationRequiredStake } from '../lib/config'
 import { useNationBalance } from '../lib/nation-token'
 import { NumberType, transformNumber } from '../lib/numbers'
-import { usePassportExpirationDate } from '../lib/passport-expiration'
+import { usePassportExpirationDate } from '../lib/passport-expiration-hook'
 import { useClaimPassport, useHasPassport } from '../lib/passport-nft'
 import { storeSignature, useSignAgreement } from '../lib/sign-agreement'
 import { useAccount } from '../lib/use-wagmi'
